requirements.txt
click==7.1.1
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.11.1
markupsafe==1.1.1
werkzeug==1.0.0
flask_cors==3.0.0
gunicorn==20.0.4
pyodbc==4.0.30
SQLAlchemy==1.3.16
PyJWT==1.7.1